Manoj_Project/fullstack_app/backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
sqlalchemy
pytesseract
pillow
